connect client no longer requires typing 'client'.
The connect command consists of a command execution framwwork (the
connect command itself) and a number of subcommands, each implemented
in a separate .sh or .py file.  The connect client is really client.py.

When it runs, the connect command reads its configuration file(s), and
if it has the following:

	[connect]
	client = true

then it's in client mode.  In client mode, the client subcommand is
_assumed_ unless the user has also used the --full option. Thus:

connect			- client mode
connect --full  - full mode
connect client  - client mode

Even in client mode, the 'connect client' syntax works.