Fix some $HOME issues in server.
Server mode chdirs to the --repo location and sets $HOME to it. Since
setup in particular requires $HOME to be normal, we need a way around
that. rcmd() now takes a keyword param userepo=<boolean>. Default is
True; if set to False, then we do not pass --repo to the server and
HOME is left intact.